Move shared prefs extensions to separate file
